Fixed: DTO's and Validation

+++ src/restaurants/restaurants.service.ts
@@ -1,7 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Restaurant } from './schemas/restaurants.schema';
-import { Model } from 'mongoose';
+import mongoose, { Model } from 'mongoose';
 import { Query } from 'src/utils/type';
 
 @Injectable()
@@ -34,12 +38,16 @@ export class RestaurantService {
   }
 
   async findOne(id: string): Promise<Restaurant> {
+    const isValid = mongoose.isValidObjectId(id);
+    if (!isValid) throw new BadRequestException('Invalid MongoDB id');
     const restaurant = await this.restuarantModel.findById(id);
     if (!restaurant) throw new NotFoundException('Restaurant not found');
     return restaurant;
   }
 
   async update(id: string, restaurant: Restaurant): Promise<Restaurant> {
+    const isValid = mongoose.isValidObjectId(id);
+    if (!isValid) throw new BadRequestException('Invalid MongoDB id');
     const restaurantToUpdate = await this.restuarantModel.findById(id);
     if (!restaurantToUpdate)
       throw new NotFoundException('Restaurant not found');
@@ -47,6 +55,8 @@ export class RestaurantService {
   }
 
   async delete(id: string): Promise<Restaurant> {
+    const isValid = mongoose.isValidObjectId(id);
+    if (!isValid) throw new BadRequestException('Invalid MongoDB id');
     const restaurant = await this.restuarantModel.findById(id);
     if (!restaurant) throw new NotFoundException('Restaurant not found');
     return this.restuarantModel.findByIdAndDelete(id);
